+++ libs/langchain/langchain/agents/agent_toolkits/__init__.py
@@ -1,6 +1,6 @@
 """Agent toolkits."""
 from langchain.agents.agent_toolkits.amadeus.toolkit import AmadeusToolkit
-from langchain.agents.agent_toolkits.azure_cognitive_services.toolkit import (
+from langchain.agents.agent_toolkits.azure_cognitive_services import (
     AzureCognitiveServicesToolkit,
 )
 from langchain.agents.agent_toolkits.csv.base import create_csv_agent
@@ -40,33 +40,33 @@ from langchain.agents.agent_toolkits.zapier.toolkit import ZapierToolkit
 
 __all__ = [
     "AmadeusToolkit",
+    "AzureCognitiveServicesToolkit",
+    "FileManagementToolkit",
+    "GmailToolkit",
+    "JiraToolkit",
+    "JsonToolkit",
+    "NLAToolkit",
+    "O365Toolkit",
+    "OpenAPIToolkit",
+    "PlayWrightBrowserToolkit",
+    "PowerBIToolkit",
+    "SQLDatabaseToolkit",
+    "SparkSQLToolkit",
+    "VectorStoreInfo",
+    "VectorStoreRouterToolkit",
+    "VectorStoreToolkit",
+    "ZapierToolkit",
+    "create_csv_agent",
     "create_json_agent",
-    "create_sql_agent",
     "create_openapi_agent",
+    "create_pandas_dataframe_agent",
     "create_pbi_agent",
     "create_pbi_chat_agent",
     "create_python_agent",
-    "create_vectorstore_agent",
-    "JsonToolkit",
-    "SQLDatabaseToolkit",
-    "SparkSQLToolkit",
-    "NLAToolkit",
-    "PowerBIToolkit",
-    "OpenAPIToolkit",
-    "VectorStoreToolkit",
-    "create_vectorstore_router_agent",
-    "VectorStoreInfo",
-    "VectorStoreRouterToolkit",
-    "create_pandas_dataframe_agent",
     "create_spark_dataframe_agent",
     "create_spark_sql_agent",
-    "create_csv_agent",
-    "ZapierToolkit",
-    "GmailToolkit",
-    "JiraToolkit",
-    "FileManagementToolkit",
-    "PlayWrightBrowserToolkit",
-    "AzureCognitiveServicesToolkit",
-    "O365Toolkit",
+    "create_sql_agent",
+    "create_vectorstore_agent",
+    "create_vectorstore_router_agent",
     "create_xorbits_agent",
 ]
