+++ libs/langchain/langchain/agents/agent_toolkits/azure_cognitive_services.py
