--- libs/langchain/langchain/agents/agent_toolkits/azure_cognitive_services/__init__.py
@@ -1,7 +1,0 @@
-"""Azure Cognitive Services Toolkit."""
-
-from langchain.agents.agent_toolkits.azure_cognitive_services.toolkit import (
-    AzureCognitiveServicesToolkit,
-)
-
-__all__ = ["AzureCognitiveServicesToolkit"]
